Update SVG path in tests and theme properties
The SVG element's path used in the tests has been updated to reflect changes in its structure. The 'ensureOrder' function in the SortingTest.kt file now checks the correct svg path "/svg/g/g/a/text" instead of "/svg/g/a/text". This was necessary because an additional group tag was added to the SVG structure to improve rendering.

Furthermore, in ButtonRenderImplTest.kt, the theme's scale property is now being set to improve visual fidelity. Additionally, the number of columns has been modified for better layout organization.