Refactor PillButtonRenderer to use abstraction
The 'PillButtonRenderer' class was refactored to extend an 'AbstractButtonRenderer' class. This is done to encapsulate common properties and methods in the 'AbstractButtonRenderer' super class, thereby promoting code-reusability and maintainability. The 'Theme' class was also updated with a new property 'panelOpacity'. Fixed hard-coded values in 'ButtonCardRendererNext'. These changes make the code easier to understand and modify in the future.