Remove class from SVG rect element definition
Removed the "raise" class from the rectangle element within SlimCardRendererNext. The class was placed awkwardly within the SVG element and was causing rendering issues. The class "raise" was not defined anywhere in the style sheet leading to ignored CSS properties. Thus, it was removed for better SVG image rendering. Enhanced the SVG visualizations whilst maintaining the same functionality.