Add filter, gradient to SlimCardRenderer's path
Added filter='buttonBlur' and fill='overlayGrad' to path tag in SlimCardRendererNext.kt as a part of UI enhancement for cards in docops. This is to give gradient effect and to improve the button visibility. Additionally, added rect tags to simulate the 3D effect on the top and bottom of the button.