Remove redundant ButtonCardRenderer, LargeCard and SlimCardRenderer
These classes were removed due to reduction in complexity and redundant code in the application.
This change helps to streamline the button rendering process and improve the maintainability of the application by
utilizing a more generic solution for button rendering.
The removal of these classes has no adverse impact on application functionality.