Refactored SVG tag generation and fixed CSS syntax
This commit lands two changes. Firstly, it adds xml escaping for link labels in the `<tspan>` SVG tag generation inside RectangleCardRenderer.kt. This is important to prevent any possible XML injection if the link label contains XML special characters.

Additionally, a syntax correction is made in the CSS styling of RoundButtonItemRenderer.kt. A semicolon replaces a comma to properly separate the 'stroke-width' and 'stroke' properties. This will ensure the hover effect takes place as expected.